+++ rocket-backend/internal/server/auth_handlers.go
@@ -1,0 +1,39 @@
+package server
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/google/uuid"
+	"rocket-backend/internal/custom_error"
+)
+
+func (s *Server) AuthHelloHandler(c *gin.Context) {
+	userID, exists := c.Get("userID")
+	if !exists {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "User not authenticated"})
+		return
+	}
+
+	userUUID, err := uuid.Parse(userID.(string))
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid user ID format"})
+		return
+	}
+
+	cred, err := s.db.GetUserByID(userUUID)
+	if err != nil {
+		if err == custom_error.ErrUserNotFound {
+			c.JSON(http.StatusNotFound, gin.H{"error": "User not found"})
+		} else {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to retrieve user"})
+		}
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "Hello from auth", "user": cred.Username})
+}
+
+func (s *Server) Authenticated(c *gin.Context) {
+	c.JSON(http.StatusOK, gin.H{"authenticated": "true"})
+}
