+++ rocket-backend/internal/server/public_handlers.go
@@ -121,20 +121,5 @@ func (s *Server) RegisterHandler(c *gin.Context) {
 		return
 	}
 
-
 	c.JSON(http.StatusOK, gin.H{"message": "User registered successfully"})
-}
-
-func (s *Server) GetUserRanking(c *gin.Context) {
-	ranking, err := s.db.GetTopUsers(100)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Error fetching ranking"})
-		return
-	}
-	if ranking == nil {
-		c.JSON(http.StatusNotFound, gin.H{"error": "Ranking for user not found"})
-		return
-	}
-
-	c.JSON(http.StatusOK, ranking)
 }
