+++ rocket-backend/internal/server/routes.go
@@ -37,11 +37,10 @@ func (s *Server) RegisterRoutes() http.Handler {
 			protected.POST("/user/statistics", s.getUserStatistics)
 			protected.POST("/user/image", s.GetUserImage)
 			protected.GET("/user/rocketpoints", s.GetRocketPoints)
+			protected.GET("/users", s.GetAllUsers)
 
 			protected.GET("/challenges/new", s.GetDailyChallenges)
 			protected.POST("/challenges/complete", s.CompleteChallenge)
-
-			protected.GET("/users", s.GetAllUsers)
 
 			protected.GET("/ranking/users", s.GetUserRanking)
 			protected.GET("/ranking/friends", s.GetFriendsRanked)
